requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
astor==0.8.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
Babel==2.14.0
bce-python-sdk==0.9.5
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.9
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
filelock==3.13.3
fire==0.6.0
Flask==3.0.2
flask-babel==4.0.0
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
imageio==2.34.0
imgaug==0.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai==0.28.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==24.0
paddleocr==2.7.0.3
paddlepaddle==2.6.1
pandas==2.2.1
pdf2docx==0.5.8
pillow==10.2.0
premailer==3.10.0
protobuf==5.26.0
psutil==5.9.8
pyclipper==1.3.0.post5
pycryptodome==3.20.0
PyMuPDF==1.20.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.2
rarfile==4.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.12
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2024.2.12
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
visualdl==2.5.3
Werkzeug==3.0.1
yarl==1.9.4
